Include library headers in .cpp
