+++ code/espurna/alexa.cpp
@@ -19,6 +19,9 @@ Copyright (C) 2016-2019 by Xose Pérez <xose dot perez at gmail dot com>
 #include "web.h"
 #include "ws.h"
 
+#include <fauxmoESP.h>
+#include <ArduinoJson.h>
+
 struct alexa_queue_element_t {
     unsigned char device_id;
     bool state;
