+++ code/espurna/influxdb.h
@@ -10,8 +10,6 @@ Copyright (C) 2017-2019 by Xose Pérez <xose dot perez at gmail dot com>
 
 #if INFLUXDB_SUPPORT
 
-#include <ESPAsyncTCP.h>
-
 bool idbSend(const char * topic, unsigned char id, const char * payload);
 bool idbSend(const char * topic, const char * payload);
 bool idbEnabled();
