+++ code/espurna/libs/AsyncClientHelpers.h
@@ -4,8 +4,6 @@
 
 #pragma once
 
-#include <ESPAsyncTCP.h>
-
 enum class AsyncClientState {
     Disconnected,
     Connecting,
